Move snippets breadcrumb items definition to generic views
